+++ src/main/java/com/hhpl/concertreserve/domain/concert/type/ReservationStatus.java
@@ -1,4 +1,4 @@
-package com.hhpl.concertreserve.domain.concert;
+package com.hhpl.concertreserve.domain.concert.type;
 
 public enum ReservationStatus {
     COMPLETE, TEMPORARY, CANCEL
