+++ src/main/java/com/hhpl/concertreserve/domain/waitingqueue/WaitingQueueScheduler.java
@@ -4,8 +4,6 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 
-import java.util.List;
-
 @Service
 @RequiredArgsConstructor
 public class WaitingQueueScheduler {
