+++ src/main/java/com/hhpl/concertreserve/domain/waitingqueue/type/WaitingQueueStatus.java
@@ -1,4 +1,4 @@
-package com.hhpl.concertreserve.domain.waitingqueue;
+package com.hhpl.concertreserve.domain.waitingqueue.type;
 
 public enum WaitingQueueStatus {
     ACTIVE, INACTIVE, EXPIRED
