+++ src/main/java/com/hhpl/concertreserve/infra/payment/PaymentJpaRepository.java
@@ -1,6 +1,6 @@
 package com.hhpl.concertreserve.infra.payment;
 
-import com.hhpl.concertreserve.domain.payment.Payment;
+import com.hhpl.concertreserve.domain.payment.model.Payment;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PaymentJpaRepository extends JpaRepository<Payment,Long> {
